refactor: 도메인별 hook, api 정리
