+++ src/features/explore/hooks/index.ts
@@ -1,0 +1,6 @@
+export { useCategory } from './useCategory';
+export {
+  useCategoryLists,
+  useCategoryTagLists,
+  useFilteredPlaces,
+} from './queries';
