+++ src/features/explore/hooks/useCategory.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { CategoryContext } from '../context/categoryContext';
+import { CategoryContext } from '../../../context/categoryContext';
 
 export const useCategory = () => {
   const context = useContext(CategoryContext);
