+++ src/features/main/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useHotPlaces } from './queries';
